Fix banners being white on the search-page
Adding the seemingly new classnames for the banner on the search-page. Seems to have been changed on the news-and-videos-page, so I just took the changes from there, and put them here. Kept the old changes for the banners as it seems that it may still be using the old classnames on the current full-release as I've understood it.